moved Score dataclass from models.py to score.py
